+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Hexadecimal (`0x...`) and binary (`0b...`) integer literal notation.
+- Support for partial type definitions.
 
 ## [0.1.0] - 2025-10-20